refactor: 위치 검색 컴포넌트 네이밍 변경
