+++ apps/web/src/components/search-place-bottom-sheet/search-list-item/style.css.ts
